find_minimal_splitting_sets if only one quorum
If there is only one (minimal) quorum, we might as well return that
quorum as the only splitting set.
This ensures correct results in cases with multiple quorums but only one
minimal quorum.
This also makes sense if there is only one quorum in the FBAS: if the
whole quorum is faulty, it can fail (and split itself) in arbitrary
ways.