Enhance symmetry detection
Workaround for detecting symmetric clusters and symmetric top tiers in
systems where a node's quorum set might not include the node itself.
According to our crawls, this is currently the case in MobileCoin. The
workaround is to include the "owner" node before doing symmetry
analyses. Such a change preserves the semantics of patched node + quorum
set combinations entirely.