Perf: leverage symmetry for skipping branches
We now identify "symmetric nodes": nodes that are a equivalent in terms
of their role in the FBAS, i.e., groups of nodes for which shuffling the
IDs wouldn't change the current analysis' results in any way. We
require that each group of symmetric nodes is used in branching
algorithms in a predetermined order - this way we can skip redundant
branches. We just have to "expand" the symmetric nodes again at the end.