Perf: pruning condition for splitting sets search
Another improvement to the `has_potential` pruning condition for
`find_minimal_splitting_sets`. Increases the speed of splitting sets
search in FBASs with many non-core splitting sets. Analyses on the core
of an FBAS (`.to_core()`) shouldn't notice any real difference.

The downside is that the pruning condition is now slightly more complex.